// commerce-admin-console/src/app/rest/models/com-ibm-commerce-approval-beans-order-approval-status-list-data-bean-_ibm_store-_summary-result-list.ts
/* tslint:disable */
export interface ComIbmCommerceApprovalBeansOrderApprovalStatusListDataBean_IBM_Store_SummaryResultList {
  submitterId?: string;
  approvalStatusId?: string;
  submitTime?: string;
  flowTypeId?: string;
  stateId?: string;
  comment?: string;
  approverGroupId?: string;
  flowId?: string;
  approveTime?: string;
  status?: string;
  approverId?: string;
  entityId?: string;
}